# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Telemt Panel API
  version: 2.0.0-draft1
  description: >
    Admin API of Telemt Panel v2 plus the unauthenticated subscription page.
    Telemt-shaped payloads follow the upstream source catalog (specs/07-telemt-sdk.md).
servers:
  - url: /
    description: Same origin; base_path is prepended by the SPA runtime when configured.
...
security:
  - session: []
paths:
...
  /api/auth/sessions:
    get:
      tags: [auth]
      operationId: listSessions
      description: Active panel sessions, newest first with the current session pinned. Cursor-paged and searchable by browser/User-Agent or IP address.
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 30 } }
        - { name: cursor, in: query, schema: { type: string }, description: Opaque cursor returned by the previous page }
        - { name: q, in: query, schema: { type: string, maxLength: 200 }, description: Case-insensitive User-Agent or IP substring }
      responses:
        "403": { description: "Authentication disabled, unapproved anonymous host, or cross-site request", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "200":
          description: Page of active sessions (current one flagged)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
    delete:
      tags: [auth]
      operationId: revokeOtherSessions
      description: Revoke all sessions except the current one.
      responses:
        "204": { description: Revoked }
        "403": { description: Authentication disabled or request rejected, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: panel_session
...
  responses:
...
    BadRequest:
      description: Invalid input
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - bad_request
            - tls_invalid_candidate
            - tls_manual_required
            - tls_prepare_busy
            - tls_prepare_unavailable
            - tls_listener_unavailable
            - tls_challenge_unavailable
            - tls_cache_unavailable
            - tls_certificate_invalid
            - tls_certificate_untrusted
            - tls_acquisition_failed
            - tls_prepare_failed
            - tls_prepare_timeout
            - tls_config_changed
            - tls_receipt_invalid
            - tls_save_failed
            - tls_restart_not_pending
            - tls_restart_failed
            - conflict
            - confirmation_required
            - invalid_credentials
            - rate_limited
            - session_expired
            - csrf_rejected
            - auth_disabled
            - quota_reset_busy
            - invalid_quota_schedule
            - quota_schedule_conflict
            - quota_schedule_storage
            - quota_schedule_unavailable
            - quota_confirmation_expired
            - quota_operation_unavailable
            - internal_error
            - branding_invalid_title
            - branding_invalid_mode
            - branding_invalid_path
            - branding_logo_unreadable
            - branding_invalid_image
            - not_found
            - telemt_unreachable
            - capability_absent
            - capability_unavailable
            - manual_restart_required
            - manual_service_control_required
            - service_action_unconfirmed
            - update_locked
            - update_version_unsupported
            - sublink_unavailable
            - log_tail_unavailable
            - log_stream_unavailable
            - log_source_error
            - invalid_webauthn_origin
            - invalid_webauthn_challenge
            - invalid_webauthn_response
            - webauthn_credential_exists
            - passkey_unavailable
            - telemt_auth_failed
            - user_exists
            - last_user_forbidden
            - read_only
            - revision_conflict
            - invalid_toml
            - invalid_config_path
            - config_unset_unsupported
            - no_changes
            - toml_projection_failed
            - reload_in_progress
            - reload_not_found
            - ambiguous_listeners
            - access_not_editable
            - section_not_editable
            - field_not_editable
            - unauthorized
            - forbidden
            - method_not_allowed
            - config_patch_not_atomic
            - payload_too_large
            - api_disabled
            - maestro_unavailable
            - unsupported_media_type
            - web_runtime_unavailable
            - web_snapshot_busy
            - web_runtime_mismatch
            - web_issuance_enabled
            - web_operation_in_progress
            - web_session_not_found
            - web_operation_not_found
            - web_vhost_not_found
            - web_profile_required
          description: >
            Machine code. Panel codes actually emitted today (grepped from every
            WriteError call site): bad_request, invalid_credentials, rate_limited,
            session_expired, csrf_rejected, auth_disabled, internal_error, not_found,
            telemt_unreachable, capability_absent, capability_unavailable,
            manual_restart_required, update_locked, sublink_unavailable,
            log_tail_unavailable, log_stream_unavailable, log_source_error,
            invalid_webauthn_origin,
            invalid_webauthn_challenge, invalid_webauthn_response,
            webauthn_credential_exists, passkey_unavailable,
            quota_reset_busy, quota_confirmation_expired, quota_operation_unavailable,
            invalid_toml, invalid_config_path, config_unset_unsupported,
            no_changes, toml_projection_failed.
            capability_absent (501) vs capability_unavailable (503) are
            deliberately distinct, not aliases: capability_absent means the
            route itself doesn't exist on this Telemt build (a bare 404/405
            with no error envelope — detected reactively, after attempting
            the call: rotate-secret, enable/disable, POST /api/telemt/reload,
            GET /api/telemt/reload/{id}); capability_unavailable means the
            route exists but the feature behind it is switched off on this
            Telemt — either known up front from the SDK's cached Capabilities
            probe (GET/PATCH /api/telemt/config, config_api) or reported by
            the response itself (GET /api/telemt/tls-fingerprints, whose
            enabled:false means runtime_edge_enabled is off; read from the
            response rather than probed so an unreachable Telemt still maps
            to 502 telemt_unreachable).
            Reserved for milestones not yet implemented: telemt_auth_failed
            (superseded on /api/telemt/info by a reachable:false
            body, not an error status — kept here for /api/telemt/config, M3).
            A well-formed Telemt *APIError whose status is 4xx and isn't otherwise
            mapped above is passed through verbatim with Telemt's own code — notably
            user_exists, last_user_forbidden, read_only, revision_conflict,
            reload_in_progress, reload_not_found, ambiguous_listeners (the latter two
            absent from Telemt's own documented error-code table but confirmed against
            its source, M3), plus any other code in Telemt's own set (07-telemt-sdk.md):
            bad_request, access_not_editable, section_not_editable, field_not_editable,
            unauthorized, forbidden, method_not_allowed, config_patch_not_atomic,
            payload_too_large, api_disabled, maestro_unavailable — except
            access_not_editable/section_not_editable/field_not_editable/
            config_patch_not_atomic/ambiguous_listeners on PATCH /api/telemt/config,
            which the panel remaps to HTTP 422 regardless of Telemt's own status.
            The WEB group (Telemt >= 3.5.3, internal/telemt/types_web.go) adds
            web_runtime_mismatch, web_issuance_enabled, web_operation_in_progress,
            web_snapshot_busy, web_session_not_found, web_operation_not_found and
            unsupported_media_type; web_runtime_unavailable is listed because it is
            Telemt's own code, but the panel remaps it to capability_unavailable
            (rule R5) so the closed-capability gate is drawn instead of an error.
            Every code in this enum must carry a message in BOTH dictionaries —
            web/src/i18n/i18n.test.ts walks this list.
        message: { type: string }
...
    SessionInfo:
      type: object
      required: [id, created, last_seen, current]
      properties:
        id: { type: string }
        created: { type: string, format: date-time }
        last_seen: { type: string, format: date-time }
        ip: { type: string }
        user_agent_label: { type: string, description: Parsed, human-readable ("iPhone · Safari") }
        auth_method: { type: string, enum: [password, passkey] }
        current: { type: boolean }
    SessionPage:
      type: object
      required: [items, total, device_count]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/SessionInfo" }
        total: { type: integer, minimum: 0 }
        device_count: { type: integer, minimum: 0, description: Distinct User-Agent and IP pairs in the filtered result }
        next_cursor: { type: string, description: Opaque cursor for the next page; absent on the last page }
...